Create README.md
This is what JetBrains wants the code to be able to do.

As far as I can tell, my code does all of these things, except Example 3. My code does return an error, but I don't know how to get it to return "Incorrect parameters." I think this might be the issue.